Upgrade gradle version to 3.0.1
In this we are also removing the androidBuildToolsVersion. When upgrading
gradle, this warning was given:

>Warning:The specified Android SDK Build Tools version (26.0.0) is ignored, as
it is below the minimum supported version (26.0.2) for Android Gradle Plugin
3.0.1.  Android SDK Build Tools 26.0.2 will be used. To suppress this warning,
remove "buildToolsVersion '26.0.0'" from your build.gradle file, as each
version of the Android Gradle Plugin now has a default version of the build
tools.

Because the version of build tools now has a default, it doesn't need to be
specified.